* Array.m ([Array_c -describe:]): Removed cast of pointer to unsigned long.
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@1419 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/collections/Array.m
@@ -331,11 +331,11 @@ PHASE(Using)
   [super describe: outputCharStream];
   if ( getBit( bits, Bit_MemberAlloc ) ) {
     sprintf( buffer, "> external member allocation at: " PTRFMT "\n",
-             (unsigned long)block );
+             block );
     [outputCharStream catC: buffer];
   } else if ( getBit( bits, Bit_DefaultMember ) ) {
     sprintf( buffer, "> default member value: " PTRFMT "\n",
-             (unsigned long)block[count] );
+             block[count] );
     [outputCharStream catC: buffer];
   }
 }
